ender-5: Initial configuration
I started from the example configuration for the Ender 5 and then
changed settings which didn't match with the stock 1.1.6 configuration.

Most notable, the x/y axis homing to min and having their directions
inverted.  There is nothing wrong with how the example was configured
(which puts the front of the bed at the front instead of the back)
but I wanted to keep the orientation the same to make comparison
with the stock firmware easier.

Other changes include feedrates, acceleration, bed size, etc which
generally match Creality's setting (except for the extruder max
feedrate of 50mm/sec instead of 25mm/sec).